% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChunkSize.R
\name{getChunkSize}
\alias{getChunkSize}
\title{Calculate the number of time steps to read for each model component.}
\usage{
getChunkSize(start_date, end_date, recordingTimeStep)
}
\arguments{
\item{start_date}{Lubridate datetime of the start of the simulation}

\item{end_date}{Lubridate datetime of the end of the simulation}

\item{recordingTimeStep}{The simulations recording time step (in seconds) in character or as numeric.}
}
\value{
A numeric of the number of time steps, including one header line, to be read.
}
\description{
RSMinerve result files can be stored in .dst files where time series of each
model component are stored in rows. \code{getChunkSize} can be used to
determine the chunk size of each model output.
}
\examples{
start_date <- lubridate::as_datetime("01.20.2021 00:00:00",
                                     format = "\%m.\%d.\%Y \%H:\%M:\%S")
end_date <- lubridate::as_datetime("01.25.2021 00:00:00",
                                   format = "\%m.\%d.\%Y \%H:\%M:\%S")
recordingTimeStep <- 3600  # In seconds
chunk_size <- getChunkSize(start_date, end_date, recordingTimeStep)
}
