% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_block.R
\name{parse_block}
\alias{parse_block}
\title{Parse individual blocks of data from lines}
\usage{
parse_block(block_lines)
}
\arguments{
\item{block_lines}{A bunch of lines read from an RSMinerve .dst simulation output file.}
}
\value{
A tibble with the data from the block.
}
\description{
Parse individual blocks of data from lines
}
