% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRSMParameters.R
\name{parseComparator}
\alias{parseComparator}
\title{Parse comparator object}
\usage{
parseComparator(line)
}
\arguments{
\item{line}{A line string to be parsed}
}
\value{
tibble with object name, parameter names and values
}
\description{
Parse comparator object
}
\keyword{internal}
