% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRSMParameters.R
\name{parseGSM}
\alias{parseGSM}
\title{Parse GSM object}
\usage{
parseGSM(line)
}
\arguments{
\item{line}{A line string to be parsed}
}
\value{
tibble with object name, parameter names and values
}
\description{
Parse GSM object
}
\keyword{internal}
