% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRSMParameters.R
\name{parseHBV}
\alias{parseHBV}
\title{Parse HBV object}
\usage{
parseHBV(line)
}
\arguments{
\item{line}{A line string to be parsed}
}
\value{
tibble with object name, parameter names and values
}
\description{
Parse HBV object
}
\keyword{internal}
