% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRSMParameters.R
\name{parseJunction}
\alias{parseJunction}
\title{Parse junction object}
\usage{
parseJunction(line)
}
\arguments{
\item{line}{A line string to be parsed}
}
\value{
tibble with object name, parameter names and values
}
\description{
Parse junction object
}
\keyword{internal}
