% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRSMParameters.R
\name{parseKinematic}
\alias{parseKinematic}
\title{Parse Kinematic object}
\usage{
parseKinematic(line)
}
\arguments{
\item{line}{A line string to be parsed}
}
\value{
tibble with object name, parameter names and values
}
\description{
Parse Kinematic object
}
\keyword{internal}
