% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRSMParameters.R
\name{parseObject}
\alias{parseObject}
\title{Parses a line with the appropriate function}
\usage{
parseObject(Object, line)
}
\arguments{
\item{Object}{Character string describing the object}

\item{line}{Character string with line to read}
}
\value{
Tibble containing parameter values read from line
}
\description{
Parses a line with the appropriate function
}
\keyword{internal}
