% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRSMParameters.R
\name{parseSOCONT}
\alias{parseSOCONT}
\title{Parse SOCONT object}
\usage{
parseSOCONT(line)
}
\arguments{
\item{line}{A line string to be parsed}
}
\value{
tibble with object name, parameter names and values
}
\description{
Parse SOCONT object
}
\keyword{internal}
