% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRSMParameters.R
\name{parseSource}
\alias{parseSource}
\title{Parse source object}
\usage{
parseSource(line)
}
\arguments{
\item{line}{A line string to be parsed}
}
\value{
tibble with object name, parameter names and values
}
\description{
Parse source object
}
\keyword{internal}
