% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRSMParameters.R
\name{parseStation}
\alias{parseStation}
\title{Parse Station object}
\usage{
parseStation(line)
}
\arguments{
\item{line}{A line string to be parsed}
}
\value{
tibble with object name, parameter names and values
}
\description{
Parse Station object
}
\keyword{internal}
