% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_minerve_forcing_csv.R
\name{read_minerve_forcing_csv}
\alias{read_minerve_forcing_csv}
\title{Reads climate and discharge data into a tibble.}
\usage{
read_minerve_forcing_csv(filename)
}
\arguments{
\item{filename}{Path to file with input data for RS Minerve.}
}
\value{
Returns a tibble of the same format as \code{data} with data
in hourly (climate) to decadal or monthly (discharge) time steps.
Includes all attributes of the csv file
}
\description{
The data is in a csv file that can also be imported to RS Minerve to form a database.
}
