% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDBCSV.R
\name{readDBCSV}
\alias{readDBCSV}
\title{Reads a RSMinerve data base csv into a tibble.}
\usage{
readDBCSV(filename, tz = "UTC")
}
\arguments{
\item{filename}{Path to file with the db data from RS Minerve.}

\item{tz}{Optional time zone string, passed to lubridate::as_datetime (for
details refer to POSIXt documentation). Defauls to "UTC"}
}
\value{
Returns a tibble of the same format as \code{data} with data
in hourly (climate) to decadal or monthly (discharge) time steps.
Includes all attributes of the csv file
}
\description{
The data is in a csv file that exported from the RS Minerve data base tab.
}
\note{
An example forcing csv file is given under \url{http://raw.githubusercontent.com/hydrosolutions/RSMinerveR/main/tests/testthat/test_translateCSVtoDST.csv}
}
