% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readForcingSTR.R
\name{readForcingSTR}
\alias{readForcingSTR}
\title{Reads climate and discharge data from a character string into a tibble.}
\usage{
readForcingSTR(string)
}
\arguments{
\item{string}{Character string with input data for RS Minerve.}
}
\value{
Returns a tibble of the same format as \code{data} with data
in hourly (climate) to decadal or monthly (discharge) time steps.
Includes all attributes of the csv file.
}
\description{
The data is in a string in the same format as the RSMinerve forcing csv
that can also be imported to RS Minerve to form a database.
}
