% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readResultDXT.R
\name{readResultDXT}
\alias{readResultDXT}
\title{Reading RSMinerve result file}
\usage{
readResultDXT(filepath)
}
\arguments{
\item{filepath}{String with path to file to be read.}
}
\description{
Reading RSMinerve result file
}
