% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRSMParameters.R
\name{readRSMParameters}
\alias{readRSMParameters}
\title{Reads parameters from RSMinerve parameter file}
\usage{
readRSMParameters(filepath)
}
\arguments{
\item{filepath}{Character with path to file to read.}
}
\value{
Tibble contianing all parameters read from file.
}
\description{
Reads parameters from RSMinerve parameter file
}
