% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSelectionCHK.R
\name{readSelectionCHK}
\alias{readSelectionCHK}
\title{Reads a chk file containing the variables selected for printing}
\usage{
readSelectionCHK(filepath)
}
\arguments{
\item{filepath}{Path to file to be read.}
}
\value{
A list with the name of the selection and the content of the Paths
in the chk file as a tibble.
}
\description{
RSMinerve writes result files or performs plots for a selection of variables.
This selection can be stored in and loaded to RSMInerve from a chk file. The
function \code{readSelectionCHK} reads such a file to a tibble.
}
\details{
The chk file is structured as shown in the following example: \cr  <?xml version="1.0" encoding="utf-8"?>  \\cr


<Selection>  \cr
<Name>New selection</Name>  \cr
<Path>Model Koksu\\Source QSpring\\Kichkinesay - QUp (m3/s)</Path>  \cr
<Path>Model Koksu\\Comparator Comparator 1\\QReference (m3/s)</Path>  \cr
<Path>Model Koksu\\Comparator Comparator 1\\QSimulation (m3/s)</Path>  \cr
...  \cr
The content of the Path parts is parsed to a tibble with columns Model,
Object, ID and Variable. The output tibble of the example above would be: \cr
A tibble: 94 x 3  \cr
Model       Object                     Variable    \cr
<chr>       <chr>                      <chr>             \cr
Model Koksu Source QSpring             Kichkinesay - QUp (m3/s)  \cr
Model Koksu Comparator Comparator 1    QReference (m3/s)     \cr
Model Koksu Comparator Comparator 1    QSimulation (m3/s)    \cr
...         ...                        ...  \cr
}
\examples{
\dontrun{
filepath <- normalizePath(file.path("test_selection.chk"))
selection_list <- readSelectionCHK(filepath)
selection_name <- selection_list[[1]]
selection_data <- selection_list[[2]]
}
}
