% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRSMParameters.R
\name{readSpecificLine}
\alias{readSpecificLine}
\title{Read specified line from file}
\usage{
readSpecificLine(filepath, linenumber)
}
\arguments{
\item{filepath}{Character path to file to be read}

\item{linenumber}{numeric or list of numerics indicating which line to read}
}
\value{
Character of line or list of character lines to read from file
}
\description{
Read specified line from file
}
\keyword{internal}
