% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRSMParameters.R
\name{writeRSMParameters}
\alias{writeRSMParameters}
\title{Write parameters to RS Minerve parameter file}
\usage{
writeRSMParameters(parameters, outfilepath)
}
\arguments{
\item{parameters}{Tibble with parameters for each RS Minerve object. Multiple parameter sets may be possible. See Details & Examples for more info on the column requirements.}

\item{outfilepath}{Character string with path to file to be written.}
}
\value{
NULL for success.
}
\description{
Write parameters to RS Minerve parameter file
}
\details{
\code{parameters} contains 5 columns:
\code{Object} (character) identifying the RSMinerve object type (e.g. Station). See Vignette Parameters for more information on the objects available in RSMinerveR.
\code{Name} is the user specified name of the object in the RSMinerve model (e.g. Meteo station A).
\code{Parameters} names all parameters for each object (e.g. Zone, X \link{m}, Y \link{m}, etc.). See Vignette Parameters for more information on the parameters of the available objects in RSMinerveR.
\code{Values} contains the parameter values (e.g. "A", 4500, 3000).
\verb{Parameter set} is used to differentiate between multiple parameter sets. The function will write length(unique(parameters$\verb{Parameter set})) parameter files, appending the value of Parameter set to the base file name.
}
\examples{
To be written
}
