% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readResultDXT.R
\name{parse_block}
\alias{parse_block}
\title{Parse individual blocks of data from lines}
\usage{
parse_block(block_lines)
}
\arguments{
\item{block_lines}{A bunch of lines read from an RSMinerve .dst simulation output file.}
}
\value{
A tibble with the data from the block.
}
\description{
Internal function called by readResultDXT to read individual data chunks from file.
}
\keyword{internal}
