% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readResultDST.R
\name{parse_block}
\alias{parse_block}
\title{Parse individual blocks of data from lines}
\usage{
parse_block(block_lines, tz = "UTC")
}
\arguments{
\item{block_lines}{A bunch of lines read from an RSMinerve .dst simulation
output file.}

\item{tz}{Time zone string to be passed to lubridate::as_datetime. Defaults
to "UTC".}
}
\value{
A tibble with the data from the block.
}
\description{
Internal function called by readResultDXT to read individual data chunks from
file.
}
\keyword{internal}
