% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readResultDXT.R
\name{readResultDXT}
\alias{readResultDXT}
\title{Reading RSMinerve result file}
\usage{
readResultDXT(filepath, chunk_size)
}
\arguments{
\item{filepath}{String with path to file to be read.}

\item{chunk_size}{Numeric telling how many lines to read for each model component.}
}
\value{
tibble with time series of simulation results for all model components.
}
\description{
Reading RSMinerve result file
}
\examples{
filepath <- normalizePath(file.path("Tutorial01-results.dst"))
chunk_size <- getChunkSize(
  lubridate::as_datetime("09.02.2013 00:00:00",format = "\%m.\%d.\%Y \%H:\%M:\%S"),
  lubridate::as_datetime("09.09.2013 00:00:00", format = "\%m.\%d.\%Y \%H:\%M:\%S"),
  3600
)
result <- readResultDXT(filepath, chunk_size)
}
