% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRSMParameters.R
\name{readRSMParameters}
\alias{readRSMParameters}
\title{Reads parameters from RSMinerve parameter file}
\usage{
readRSMParameters(filepath)
}
\arguments{
\item{filepath}{Character with path to file to read.}
}
\value{
Tibble containing all parameters read from file.
}
\description{
Reads parameters from RSMinerve parameter file
}
\details{
The suported RSMInerve objects are described in the vignette Overview Objects.
\code{parameters} contains 5 columns:
\code{Object} (character) identifying the RSMinerve object type (e.g. Station). See Vignette Parameters for more information on the objects available in RSMinerveR.
\code{Name} is the user specified name of the object in the RSMinerve model (e.g. Meteo station A).
\code{Zone} is the name of the zone an object is assigned to (e.g. "A").
\code{Parameters} names all parameters for each object (e.g. X \link{m}, Y \link{m}, etc.). The parameter names are the same as in the RS Minerve parameter file. See Vignette Parameters for more information on the parameters of the available objects in RSMinerveR.
\code{Values} contains the parameter values as numerics (e.g. 4500, 3000).
\verb{Parameter set} is used to differentiate between multiple parameter sets. The function will write length(unique(parameters$\verb{Parameter set})) parameter files, appending the value of Parameter set to the base file name.
}
\examples{
filepath <- normalizePath(file.path("Tutorial_Parameters.txt"))
params <- readRSMParameters(filepath)
}
