% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateCSVtoDST.R
\name{translateCSVtoDST}
\alias{translateCSVtoDST}
\title{Translate csv to dst}
\usage{
translateCSVtoDST(csvfilepath, tz = "UTC")
}
\arguments{
\item{csvfilepath}{string path to csv file to be read.}

\item{tz}{Time zone string to be passed to base::format. Defaults to "UTC".}
}
\value{
NULL for success and 1 for failure.
}
\description{
!UNDER DEVELOPMENT! This function is meant to translate forcing data for RSMinerve in the csv format to dst which can be loaded when scripting RSMinerve. The dst output file is written to the same location as the input file.
}
\details{
When scripting RSMinerve, climate forcing is read from dst files and not from csv files as are used when RSMinerve is run from the user interface. The csv files can be transformed to dst files using this function.
}
